% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{add_exp_class}
\alias{add_exp_class}
\title{Add "exp" class}
\usage{
add_exp_class(x)
}
\arguments{
\item{x}{An object, usually containing a logistic regression model}
}
\description{
This function prepares an object so that the S3 method tidy.exp is called
on it by the tidy() function (e.g., in `modelsummary::msummary`)
}
