% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaTablesSurvey.R
\name{apa.cor.table.survey}
\alias{apa.cor.table.survey}
\title{Create a correlation table from survey data with summary statistics in APA style}
\source{
Based on the apaTables \code{apa.cor.table()} function, but adapted to
accept survey correlation matrix
}
\usage{
apa.cor.table.survey(
  cor.matrix,
  filename = NA,
  table.number = NA,
  landscape = TRUE,
  note = ""
)
}
\arguments{
\item{cor.matrix}{A survey correlation matrix, usually returned from \code{
survey_cor_matrix()}}

\item{filename}{(optional) Output filename document filename (must end in .rtf or .doc only)}

\item{table.number}{Integer to use in table number output line}

\item{landscape}{(TRUE/FALSE) Make RTF file landscape}

\item{note}{Additional notes to show under the table.}
}
\value{
A table that can be printed to the console. The function is typically
more useful for saving a table to file, using the respective argument.
}
\description{
This function, based on apaTables' \code{apa.cor.table()}, creates (and
optionally saves) a correlation table with summary statistics.
}
