% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tables.R
\name{apa_cor_table}
\alias{apa_cor_table}
\title{Create a correlation table with summary statistics in APA style}
\source{
Based on the apaTables \code{apa.cor.table()} function, but adapted to
accept weighted correlation matrices and work with the `gt` package instead`
}
\usage{
apa_cor_table(cor_matrix, filename = NULL, notes = list(NULL), stars = NULL)
}
\arguments{
\item{cor_matrix}{A correlation matrix, for example returned from \code{
survey_cor_matrix()}, \code{wtd_cor_matrix()}}

\item{filename}{the file name to create on disk. Include '.html' extension to best preserve formatting (see gt::gtsave for details)}

\item{notes}{List of additional notes to show under the table.}

\item{stars}{A character vector to change the significance symbols (see details in `sigstars`)}
}
\value{
A table that can be printed in the RStudio console to be shown in the
viewer. Unless it is to be post-processed with further `gt` functions, it should
usually be saved by passing a filename argument.
}
\description{
This function, creates (and #' optionally saves) a correlation table with
summary statistics. It accepts
}
