% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{corstars}
\alias{corstars}
\title{Calculates correlation matrix with significance stars and descriptives}
\source{
Adapted from
 http://www.sthda.com/english/wiki/elegant-correlation-table-using-xtable-r-package
}
\usage{
corstars(
  x,
  method = c("pearson", "spearman", "kendall"),
  remove_triangle = c("upper", "lower"),
  desc_col = T,
  adjust = "none",
  ...
)
}
\arguments{
\item{x}{Dataframe of variables that can be coerced to numeric.}

\item{method}{method="pearson" is the default value.  The alternatives to be passed to cor are "spearman" and "kendall"}

\item{remove_triangle}{Given that the upper and lower triangles of the
correlation matrix contain the same information, either should be dropped.
This parameter specific which.}

\item{desc_col}{Logical value that indicates whether a column with means
and standard deviations should be added in front of the correlations}

\item{adjust}{What adjustment for multiple tests should be used? ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"). See \code{\link{p.adjust}} for details about why to use "holm" rather than "bonferroni"). }

\item{...}{
  Arguments passed on to \code{\link[=sigstars]{sigstars}}
  \describe{
    \item{\code{stars}}{A character vector to change the significance symbols (see details)}
  }}
}
\value{
A dataframe including the correlation matrix
}
\description{
Calculates the correlation matrix between a given set of variables and
optionally includes a column with means and standard deviations of the
variables.
}
\details{

}
