% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpWrappers.R
\name{.is.numeric_col}
\alias{.is.numeric_col}
\title{Tests whether a column in df, specificied by string, is numeric}
\usage{
.is.numeric_col(col, df)
}
\description{
Tests whether a column in df, specificied by string, is numeric
}
