% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{glue_warn}
\alias{glue_warn}
\title{glue::glue wrapper that warns when items evaluate to NULL}
\usage{
glue_warn(
  ...,
  .sep = "",
  .envir = parent.frame(),
  .open = "{",
  .close = "}",
  .na = "NA"
)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[glue:glue]{glue::glue}}
  \describe{
    \item{\code{.transformer}}{[\verb{function]}\cr A function taking three parameters \code{code}, \code{envir} and
\code{data} used to transform the output of each block before during or after
evaluation. For example transformers see \code{vignette("transformers")}.}
    \item{\code{.trim}}{[\code{logical(1)}: \sQuote{TRUE}]\cr Whether to trim the input
template with \code{trim()} or not.}
  }}

\item{.sep}{[\code{character(1)}: \sQuote{""}]\cr Separator used to separate elements.}

\item{.envir}{[\code{environment}: \code{parent.frame()}]\cr Environment to evaluate each expression in. Expressions are
evaluated from left to right. If \code{.x} is an environment, the expressions are
evaluated in that environment and \code{.envir} is ignored. If \code{NULL} is passed it is equivalent to \code{\link[=emptyenv]{emptyenv()}}.}

\item{.open}{\[`character(1)`: \sQuote{\\\{}]\cr The opening delimiter. Doubling the
full delimiter escapes it.}

\item{.close}{\[`character(1)`: \sQuote{\\\}}]\cr The closing delimiter. Doubling the
full delimiter escapes it.}

\item{.na}{[\code{character(1)}: \sQuote{NA}]\cr Value to replace NA values
with. If \code{NULL} missing values are propagated, that is an \code{NA} result will
cause \code{NA} output. Otherwise the value is replaced by the value of \code{.na}.}
}
\description{
glue::glue wrapper that warns when items evaluate to NULL
}
