% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaTablesSurvey.R
\name{lm_with_std}
\alias{lm_with_std}
\title{Creates a summary table comparing standardised and non-standardised linear models}
\usage{
lm_with_std(
  mod,
  std_mod,
  conf_level = 0.95,
  fmt = "\%.2f",
  statistic_vertical = FALSE,
  filename = NULL,
  ...
)
}
\arguments{
\item{mod}{A lm-model/mira object of lm models, with variables not standardised}

\item{std_mod}{A lm-model/mira object of lm models, with standardised variables. Can be
created with \code{\link{lm_std}}}

\item{conf_level}{Confidence level to use for confidence intervals, defaults to .95}

\item{fmt}{string which specifies how numeric values will be rounded. This
string is passed to the `sprintf` function. '\%.3f' will keep 3 digits after
the decimal point with trailing zero. '\%.5f' will keep 5 digits. '\%.3e' will
use exponential notation. See `?sprintf` for more options.}

\item{statistic_vertical}{TRUE if statistics should be printed below
estimates. FALSE if statistics should be printed beside estimates.}

\item{filename}{the file name to create on disk. Ensure that an extension compatible with the output types is provided ('.html', '.tex', '.ltx', '.rtf'). Read '?gt::gtsave' for further details.}

\item{...}{
  Arguments passed on to \code{\link[modelsummary:modelsummary]{modelsummary::modelsummary}}
  \describe{
    \item{\code{coef_map}}{named character vector. Names refer to the original variable
names. Values refer to the variable names that will appear in the table.
Coefficients which are omitted from this vector will be omitted from the
table. The table will be ordered in the same order as this vector.}
    \item{\code{coef_omit}}{string regular expression. Omits all matching coefficients
from the table (using `stringr::str_detect`).}
    \item{\code{gof_map}}{data.frame with four columns: `raw`, `clean`, `fmt`, and
`omit`. See `modelsummary::gof_map`}
    \item{\code{gof_omit}}{string regular expression. Omits all matching gof statistics from
the table (using `stringr::str_detect`).}
    \item{\code{title}}{string}
    \item{\code{subtitle}}{string}
    \item{\code{notes}}{list of notes to append to the bottom of the table.}
    \item{\code{add_rows}}{list of character vectors, each of length equal to the number
of models + 1.}
    \item{\code{add_rows_location}}{integer or NULL. custom rows will be added to the
bottom of the table if this parameter is NULL, or after the position set by
this integer.}
  }}
}
\description{
This function creates a summary table for lm models (including mice::mira objects
containing lm-models) that shows a standardised and non-standardised version of the model
side-by-side.
}
