% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_quickpred_extension.R
\name{mice_check.dataform}
\alias{mice_check.dataform}
\title{Create mice predictorMatrix including unordered factors (extends mice::quickpred)}
\source{
mice v 3.11.4, written by Stef van Buuren.
}
\usage{
mice_check.dataform(data)
}
\description{
Create mice predictorMatrix including unordered factors (extends mice::quickpred)
}
