% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{p_pct-deprecated}
\alias{p_pct-deprecated}
\title{Formats a number to print as percentage}
\usage{
p_pct(x, digits)
}
\arguments{
\item{x}{A number}

\item{digits}{The number of digits after the percentage point. Defaults to 1. Trailing zeroes are shown as needed.}
}
\description{
Takes a number and returns it as a formatted string expressing the percentage
}
\seealso{
\code{\link{rNuggets-deprecated}}
}
\keyword{internal}
