% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{p_pct}
\alias{p_pct}
\title{Formats a number to print as percentage}
\usage{
p_pct(x, digits = 1)
}
\arguments{
\item{x}{A number}

\item{digits}{The number of digits after the percentage point. Defaults to 1}
}
\description{
Takes a number and returns it as a formatted string expressing the percentage
}
