% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_tables.R
\name{polr_with_std}
\alias{polr_with_std}
\title{Creates a summary table comparing standardised and non-standardised
proportional odd logistic regression models}
\usage{
polr_with_std(
  mod,
  std_mod,
  OR = TRUE,
  conf_level = 0.95,
  fmt = "\%.2f",
  statistic_vertical = FALSE,
  filename = NULL,
  model_names = NULL,
  show_nimp = FALSE,
  notes = list(),
  ...
)
}
\arguments{
\item{mod}{A polr-model/mira object of polr models, with variables not standardised (or a list of such models)}

\item{std_mod}{A polr-model/mira object of polr models, with standardised predictor variables (or a list of such models)}

\item{OR}{Logical. Shoulds odds ratios be shown instead of typical coefficients. If TRUE, estimates are exponentiated}

\item{conf_level}{Confidence level to use for confidence intervals, defaults to .95}

\item{fmt}{an integer, string, or function which determines how to format numeric values:
\itemize{
  \item integer: the number of digits to keep after the period (`format(round(x, fmt), nsmall=fmt)`)
  \item character: string is passed to the `sprintf` function. '\%.3f' will keep 3 digits after the decimal point with trailing zero. '\%.5f' will keep 5 digits. '\%.3e' will use exponential notation. See `?sprintf` for more options.
  \item function: a function which returns a formatted character string
}}

\item{statistic_vertical}{deprecated argument. Supply a vector of strings or `glue` strings to the `estimate` instead.}

\item{filename}{the file name to create on disk. Include '.html' extension to best preserve formatting (see gt::gtsave for details)}

\item{model_names}{If several pairs of models are to be plotted side by side, indicate the label for each *pair* here}

\item{show_nimp}{Logical. If mira objects are passed, this determines whether the number of imputations will be reported as a model statistic}

\item{notes}{List of notes to append to bottom of table. An explanation of significance stars is automatically added. A note is also added
stating that dummy variables were not scaled in standardization. If you approached standardisation differently, that should be removed.}

\item{...}{
  Arguments passed on to \code{\link[modelsummary:modelsummary]{modelsummary::modelsummary}}
  \describe{
    \item{\code{output}}{filename or object type (string)
\itemize{
  \item Supported filename extensions: .html, .tex, .md, .txt, .png, .jpg.
  \item Supported object types: "default", "html", "markdown", "latex", "data.frame", "gt", "kableExtra", "huxtable", "flextable".
  \item When a file name is supplied to the `output` argument, the table is written immediately to file. If you want to customize your table by post-processing it with functions provided by the `gt` or `kableExtra` packages, you need to choose a different output format (e.g., "gt", "latex", "html", "markdown"), and you need to save the table after post-processing using the `gt::gtsave`, `kableExtra::save_kable`, or `cat` functions.
}}
    \item{\code{coef_map}}{named character vector. Values refer to the variable names
that will appear in the table. Names refer to the original term names stored
in the model object, e.g. c("hp:mpg"="hp X mpg") for an interaction term.
Coefficients that are omitted from this vector will be omitted from the
table. The table will be ordered in the same order as this vector.}
    \item{\code{coef_omit}}{string regular expression. Omits all matching coefficients
from the table (using `grepl(perl=TRUE)`).}
    \item{\code{coef_rename}}{named character vector. Values refer to the variable names
that will appear in the table. Names refer to the original term names stored
in the model object, e.g. c("hp:mpg"="hp X mpg") for an interaction term.}
    \item{\code{gof_map}}{data.frame with four columns: `raw`, `clean`, `fmt`, and
`omit`. If `gof_map` is NULL, then `modelsummary` will use this data frame
by default: `modelsummary::gof_map` By default, all the statistics produced
by `broom::glance` will be included unless they are omitted explicitly in
`gof_map`.}
    \item{\code{gof_omit}}{string regular expression. Omits all matching gof statistics from
the table (using `grepl(perl=TRUE)`).}
    \item{\code{estimate}}{character name of the estimate to display. Must be a column
name in the data.frame produced by `tidy(model)`. In the vast majority of
cases, the default value of this argument should not be changed.}
    \item{\code{add_rows}}{a data.frame (or tibble) with the same number of columns as
your main table. By default, rows are appended to the bottom of the table.
You can define a "position" attribute of integers to set the row positions.
See examples.}
    \item{\code{title}}{string}
    \item{\code{align}}{A character string of length equal to the number of columns in
the table.  "lcr" means that the first column will be left-aligned, the 2nd
column center-aligned, and the 3rd column right-aligned.}
  }}
}
\description{
This function creates a summary table for polr models (including mice::mira objects
containing polr-models) that shows a standardised and non-standardised version of the model
side-by-side. Several pairs of such models can be compared side-by-side.
}
