% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNuggets.R
\docType{package}
\name{rNuggets-package}
\alias{rNuggets}
\alias{rNuggets-package}
\title{rNuggets: Package of useful helper functions for data analysis}
\description{
This package contains a set of functions that I have written
    to facilitate my personal data analysis workflow (mostly for research
    in social psychology and related teaching). If anyone else finds them 
    helpful, I'd be delighted and then be grateful for any bug reports.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukasWallrich/rNuggets}
  \item Report bugs at \url{https://github.com/LukasWallrich/rNuggets/issues}
}

}
\author{
\strong{Maintainer}: Lukas Wallrich \email{lukas.wallrich@gmail.com}

}
