% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNuggets.R
\docType{package}
\name{rNuggets}
\alias{rNuggets}
\alias{rNuggets-package}
\title{A set of helper functions for statistics and data analysis}
\description{
This package contains a diverse set of helper functions to support tasks that
I frequently perform as part of my research in social psychology. They are
primarily written for my own use, but might be helpful to others. As they cover
everything from statistical modeling, graphing and presentation to project management,
the package has a huge list of dependencies. Eventually, this might be split
into a few focused packages that will be more lightweight and easier to grasp.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LukasWallrich/rNuggets}
  \item Report bugs at \url{https://github.com/LukasWallrich/rNuggets/issues}
}

}
\author{
\strong{Maintainer}: Lukas Wallrich \email{lukas.wallrich@gmail.com}

}
