% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{round_}
\alias{round_}
\title{Round function that returns trailing zeroes}
\usage{
round_(x, digits = 0)
}
\arguments{
\item{x}{Numeric vector to be rounded}

\item{digits}{Number of significant digits}
}
\value{
Character vector of rounded values, with trailing zeroes as needed to show `digits` figures after the decimal point
}
\description{
Round function that returns trailing zeroes
}
