% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpWrappers.R
\name{run_lm}
\alias{run_lm}
\title{lm() for pipes - data as first argument}
\usage{
run_lm(df, formula, ...)
}
\arguments{
\item{df}{Data for modeling}

\item{formula}{an object of class \code{"\link[stats]{formula}"} (or one that
    can be coerced to that class): a symbolic description of the
    model to be fitted.  The details of model specification are given
    under \sQuote{Details}.}

\item{...}{
  Arguments passed on to \code{\link[stats:lm]{stats::lm}}
  \describe{
    \item{\code{data}}{an optional data frame, list or environment (or object
    coercible by \code{\link{as.data.frame}} to a data frame) containing
    the variables in the model.  If not found in \code{data}, the
    variables are taken from \code{environment(formula)},
    typically the environment from which \code{lm} is called.}
    \item{\code{subset}}{an optional vector specifying a subset of observations
    to be used in the fitting process.}
    \item{\code{weights}}{an optional vector of weights to be used in the fitting
    process.  Should be \code{NULL} or a numeric vector.
    If non-NULL, weighted least squares is used with weights
    \code{weights} (that is, minimizing \code{sum(w*e^2)}); otherwise
    ordinary least squares is used.  See also \sQuote{Details},}
    \item{\code{na.action}}{a function which indicates what should happen
    when the data contain \code{NA}s.  The default is set by
    the \code{na.action} setting of \code{\link{options}}, and is
    \code{\link[stats]{na.fail}} if that is unset.  The \sQuote{factory-fresh}
    default is \code{\link[stats]{na.omit}}.  Another possible value is
    \code{NULL}, no action.  Value \code{\link[stats]{na.exclude}} can be useful.}
    \item{\code{method}}{the method to be used; for fitting, currently only
    \code{method = "qr"} is supported; \code{method = "model.frame"} returns
    the model frame (the same as with \code{model = TRUE}, see below).}
    \item{\code{model}}{logicals.  If \code{TRUE} the corresponding
    components of the fit (the model frame, the model matrix, the
    response, the QR decomposition) are returned.
  }
    \item{\code{x}}{logicals.  If \code{TRUE} the corresponding
    components of the fit (the model frame, the model matrix, the
    response, the QR decomposition) are returned.
  }
    \item{\code{y}}{logicals.  If \code{TRUE} the corresponding
    components of the fit (the model frame, the model matrix, the
    response, the QR decomposition) are returned.
  }
    \item{\code{qr}}{logicals.  If \code{TRUE} the corresponding
    components of the fit (the model frame, the model matrix, the
    response, the QR decomposition) are returned.
  }
    \item{\code{singular.ok}}{logical. If \code{FALSE} (the default in S but
    not in \R) a singular fit is an error.}
    \item{\code{contrasts}}{an optional list. See the \code{contrasts.arg}
    of \code{\link[stats]{model.matrix.default}}.}
    \item{\code{offset}}{this can be used to specify an \emph{a priori} known
    component to be included in the linear predictor during fitting.
    This should be \code{NULL} or a numeric vector or matrix of extents
    matching those of the response.  One or more \code{\link[stats]{offset}} terms can be
    included in the formula instead or as well, and if more than one are
    specified their sum is used.  See \code{\link[stats]{model.offset}}.}
  }}
}
\description{
Within a dplyr-pipe, running lm() is often complicated be the placing of the
data argument. This wrapper places data first.
}
\details{
Note that the model formula is replaced by a string in the model object -
this is a quick fix for a bug that led to the fomula not being displayed
but could potentially lead to downstream issues.
}
