% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tables.R
\name{survey_cor_matrix}
\alias{survey_cor_matrix}
\title{Create a correlation matrix from survey data with summary statistics}
\usage{
survey_cor_matrix(svy_df, var_names)
}
\arguments{
\item{svy_df}{A survey object created with the survey or srvyr package. Only
numeric variables will be included in the result.}

\item{var_names}{Named character vector to rename variables to - most helpful
if return is to be passed to some print function}
}
\value{
A correlation matrix list in the format provided by
\code{jtools::svycor()} with the addition of a \code{desc}-element with means
and standard deviations of the variables.
}
\description{
This function wraps jtools::svycor() so that it works in a srvyr-pipeline,
runs bootstrapped significance-tests and calculates weighted summary
statistics. Only numeric variables are included in the result.
}
\examples{
\dontrun{
if (requireNamespace("survey") & requireNamespace("srvyr")) {
library(survey)
library(srvyr)
data(api)
dstrat <- apistrat \%>\% as_survey_design(1, strata = stype, fpc = fpc, weight = pw)

var_names <- c(meals = "Share subsidized meals", ell = "English language learners",
              growth = "Performance Change")

# Print correlation matrix
survey_cor_matrix(dstrat, var_names)
}
}
}
