% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{svy_group_means}
\alias{svy_group_means}
\title{Show group counts and group means in srvyr data}
\usage{
svy_group_means(df, gr, mean_vars, tbl_title, quietly = F)
}
\arguments{
\item{df}{A srvyr survey object}

\item{gr}{Character. The name of the grouping variable in df.}

\item{mean_vars}{Character vector. Names of one or more variables in df to calculate means for.}

\item{tbl_title}{Character. Title for summary table to be printed.}

\item{quietly}{Logical. Calculate means without displaying them?}
}
\value{
Dataframe with group counts and means
}
\description{
This function groups srvyr data by a grouping variable and then calculates
and displays group means and counts with standard errors.
}
