% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tidy.polr_p}
\alias{tidy.polr_p}
\title{Tidy polr with added p-values}
\source{
https://r.789695.n4.nabble.com/p-values-of-plor-td4668100.html
}
\usage{
\method{tidy}{polr_p}(x, ...)
}
\arguments{
\item{x}{An object containing a `polr` model.}

\item{...}{Arguments passed on to the `tidy.polr` function}
}
\description{
This function calls tidy.polr and then adds a p.value column based
on the `MASS::dropterm` chi-squared test. This approach to testing
the significance of model terms was recommeded by Prof Brian Ripley,
the author of the MASS package.
}
