% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\name{to_tribble}
\alias{to_tribble}
\title{Converts a tibble to tribble code}
\usage{
to_tribble(x, show = FALSE)
}
\arguments{
\item{x}{The tibble to be converted into tribble code}

\item{show}{Logical. Print code (otherwise, returned - print with `cat()` to get linebreaks etc)}
}
\description{
Tribbles are an easy way to legibly input data, and therefore helpful for teaching and interactive work. This function takes
a tibble and returns code that can recreate it.
}
