% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{within_km}
\alias{within_km}
\title{Identify rows of dataframe with observations in certain distance from point}
\usage{
within_km(df, start_longitude, start_latitude, km)
}
\arguments{
\item{df}{Dataframe with observations, must include latitude and longitude
variables in degrees}

\item{start_longitude}{Longitude of starting point in degrees.}

\item{start_latitude}{Latitude of starting point in degrees.}

\item{km}{Distance from starting point to filter line by}
}
\value{
Logical vector indicating which rows of \code{df} are within
 \code{km} from starting point
 @export
}
\description{
\code{within_km} tests for each line of a dataframe with longitude and
latitude variables whether the location so described is within a certain
radius from a given starting point.
}
\details{
Note that this function is very inefficient. It calculates the distance of
each line to the starting point using the Haversine formula. If you need to
use this regularly or on large datasets, a previous step to throw out lines
that are clearly too far away based on either latitude or longitude would
greatly increase efficiency.
}
