% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apaTablesSurvey.R
\name{wtd_cor_matrix_mi}
\alias{wtd_cor_matrix_mi}
\title{Create a correlation matrix from multiply imputed data with weights}
\source{
Takes some code from the \code{miceadds::micombine.cor} function,
but adapted to use weights and return in the format accepted by
\code{apa.cor.table.survey}
}
\usage{
wtd_cor_matrix_mi(mi_list, weights)
}
\arguments{
\item{mi_list}{A list of dataframes of multiple imputation results}

\item{weights}{A variable within mi_list that gives the survey weights}
}
\value{
A correlation matrix list similar to the format provided by
\code{jtools::svycor()} with the addition of a \code{desc}-element with means
and standard deviations of the variables.
}
\description{
This function takes an imputationList with a vector of weights and returns
a correlation matrix for all numeric variables as well as a list of
descriptives that pools the results across all imputations.
}
\details{
Variables starting with . are dropped, as these are likely to be .imp and .id
from mice. If you want correlations for such variables, rename them.
}
