% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{wtd_group_means_mi}
\alias{wtd_group_means_mi}
\title{Show group counts and group means in multiply imputed and weighted data}
\usage{
wtd_group_means_mi(mi_list, mean_var, gr, weights)
}
\arguments{
\item{mi_list}{A list of dataframes of multiple imputation results}

\item{mean_var}{Variable in mi_list to calculate means for.}

\item{gr}{Grouping variable in mi_list}

\item{weights}{Variable within mi_list that gives the survey weights}
}
\value{
A tiblle with means (M), standard deviations (SD) and weighted counts (N) per group
}
\description{
This function groups srvyr data by a grouping variable and then calculates
and displays group means and counts with standard errors.
}
