% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tables.R
\name{apa_cor_table}
\alias{apa_cor_table}
\title{Create a correlation table with summary statistics in APA style}
\source{
Based on the apaTables \code{apa.cor.table()} function, but adapted to
accept weighted correlation matrices and work with the `gt` package instead`
}
\usage{
apa_cor_table(
  cor_matrix,
  filename = NULL,
  notes = list(NULL),
  stars = NULL,
  add_title = FALSE,
  extras = NULL
)
}
\arguments{
\item{cor_matrix}{A correlation matrix, for example returned from
\code{cor_matrix()}, \code{survey_cor_matrix()}, or \code{wtd_cor_matrix_mi()}}

\item{filename}{the file name to create on disk. Include '.html' extension to best preserve formatting (see gt::gtsave for details)}

\item{notes}{List of additional notes to show under the table.}

\item{stars}{A character vector to change the significance symbols (see details in `sigstars`)}

\item{add_title}{Should title be added to table? Set to TRUE for default title or to character for custom title}

\item{extras}{Tibble of additional columns to be added after the descriptives column - needs to be sorted in the same order as the `desc` element in the cor_matrix}
}
\value{
A table that can be printed in the RStudio console to be shown in the
viewer. Unless it is to be post-processed with further `gt` functions, it should
usually be saved by passing a filename argument.
}
\description{
This function, creates (and optionally saves) a correlation table with
summary statistics. It accepts correlation matrices from various functions
in this package as its first argument
}
