% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_tables.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Calculates correlation matrix with significance tests and descriptives}
\source{
Adapted from
 http://www.sthda.com/english/wiki/elegant-correlation-table-using-xtable-r-package
}
\usage{
cor_matrix(
  x,
  var_names = NULL,
  method = c("pearson", "spearman", "kendall"),
  adjust = "none",
  ...
)
}
\arguments{
\item{x}{Dataframe of variables that can be coerced to numeric.}

\item{var_names}{A named character vector with the names that should be displayed
for variables. If NULL, then the variables are not renamed.}

\item{method}{method="pearson" is the default value.  The alternatives to be passed to cor are "spearman" and "kendall"}

\item{adjust}{What adjustment for multiple tests should be used? ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none"). See \code{\link{p.adjust}} for details about why to use "holm" rather than "bonferroni"). }

\item{...}{
  Arguments passed on to \code{\link[psych:corr.test]{psych::corr.test}}
  \describe{
    \item{\code{use}}{use="pairwise" is the default value and will do pairwise deletion of cases. use="complete" will select just complete cases. }
    \item{\code{alpha}}{alpha level of confidence intervals}
  }}
}
\value{
A list including the correlation matrix, p-values, standard errors, confidence intervals and descriptives
}
\description{
Calculates the correlation matrix between the numeric variables in a given dataframe and
includes descriptives (mean and standard deviation) - ready for creating a nice table with \code{\link{apa_cor_table}}
}
