% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_tables.R
\name{polr_with_std}
\alias{polr_with_std}
\title{Creates a summary table comparing standardised and non-standardised
proportional odd logistic regression models}
\usage{
polr_with_std(
  mod,
  std_mod,
  OR = TRUE,
  conf_level = 0.95,
  fmt = "\%.2f",
  statistic_vertical = FALSE,
  filename = NULL,
  model_names = NULL,
  show_nimp = FALSE,
 
    notes = list("Given that dummy variables loose their interpretability when standardised (Fox, 2015), standardised OR are only shown for continuous predictors."),
  ...
)
}
\arguments{
\item{mod}{A polr-model/mira object of polr models, with variables not standardised (or a list of such models)}

\item{std_mod}{A polr-model/mira object of polr models, with standardised predictor variables (or a list of such models)}

\item{OR}{Logical. Shoulds odds ratios be shown instead of typical coefficients. If TRUE, estimates are exponentiated}

\item{conf_level}{Confidence level to use for confidence intervals, defaults to .95}

\item{fmt}{string which specifies how numeric values will be rounded. This
string is passed to the `sprintf` function. '\%.3f' will keep 3 digits after
the decimal point with trailing zero. '\%.5f' will keep 5 digits. '\%.3e' will
use exponential notation. See `?sprintf` for more options.}

\item{statistic_vertical}{TRUE if statistics should be printed below
estimates. FALSE if statistics should be printed beside estimates.}

\item{filename}{the file name to create on disk. Include '.html' extension to best preserve formatting (see gt::gtsave for details)}

\item{model_names}{If several pairs of models are to be plotted side by side, indicate the label for each *pair* here}

\item{show_nimp}{Logical. If mira objects are passed, this determines whether the number of imputations will be reported as a model statistic}

\item{notes}{List of notes to append to bottom of table. An explanation of significance stars is automatically added. If the std models were run with a helper function in this package, a note regarding the standardisation is also automatically added.}

\item{...}{
  Arguments passed on to \code{\link[modelsummary:modelsummary]{modelsummary::modelsummary}}
  \describe{
    \item{\code{output}}{filename or object type (string)
\itemize{
  \item Supported filename extensions: .html, .tex, .md, .txt, .png, .jpg. 
  \item Supported object types: "default", "html", "markdown", "latex", "data.frame", "gt", "kableExtra", "huxtable", "flextable".
  \item When a file name is supplied to the `output` argument, the table is written immediately to file. If you want to customize your table by post-processing it with functions provided by the `gt` or `kableExtra` packages, you need to choose a different output format (e.g., "gt", "latex", "html", "markdown"), and you need to save the table after post-processing using the `gt::gtsave`, `kable::save_kable`, or `cat` functions.
}}
    \item{\code{coef_map}}{named character vector. Names refer to the original variable
names. Values refer to the variable names that will appear in the table.
Coefficients which are omitted from this vector will be omitted from the
table. The table will be ordered in the same order as this vector.}
    \item{\code{coef_omit}}{string regular expression. Omits all matching coefficients
from the table (using `stringr::str_detect`).}
    \item{\code{gof_map}}{data.frame with four columns: `raw`, `clean`, `fmt`, and
`omit`. See `modelsummary::gof_map`}
    \item{\code{gof_omit}}{string regular expression. Omits all matching gof statistics from
the table (using `stringr::str_detect`).}
    \item{\code{add_rows}}{a data.frame (or tibble) with the same number of columns as
your main table. By default, rows are appended to the bottom of the table.
You can define a "position" attribute of integers to set the row positions.
See examples.}
    \item{\code{title}}{string}
    \item{\code{estimate}}{character name of the estimate to display. Must be a column
name in the data.frame produced by `tidy(model)`. In the vast majority of
cases, the default value of this argument should not be changed.}
  }}
}
\description{
This function creates a summary table for polr models (including mice::mira objects
containing polr-models) that shows a standardised and non-standardised version of the model
side-by-side. Several pairs of such models can be compared side-by-side.
}
