% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Others.R
\encoding{UTF-8}
\name{sigstars}
\alias{sigstars}
\title{Significance stars for p-values}
\source{
Adapted from
 http://www.sthda.com/english/wiki/elegant-correlation-table-using-xtable-r-package
}
\usage{
sigstars(p, stars = NULL)
}
\arguments{
\item{p}{A \emph{p}-value or (more commonly) a vector of \emph{p}-values}

\item{stars}{A character vector to change the significance symbols (see details)}
}
\value{
A character vector of significance stars for each \emph{p}-value,
each padded with spaces to be four characters long
}
\description{
Function returns significance stars for \emph{p}-values, most likely for use
in tables that report the results of multiple statistical tests.
}
\details{
Symbols and tresholds are *** \emph{p} < .001,  ** \emph{p} < .01, * \emph{p}
 < .05 and † p < .1. The symbols can be changed by passing a 5-element
 character vector to the \code{stars} argument. For alignment, each should
 have the same length, they should indicate the signs for the four categories
 indicated above and for non-significant values (can of course be "   "). This
 argument makes most sense when marginal significance should not be indicated
 or when non-significant values should explicitly be labeled (e.g., as ns)
}
